display exception messages on errors
Noticed this using Jekyll - the exception explanations aren't in the
`serve` output.  They are helpful to have, so this adds them.

This does not touch line 75, which looks like it was wanting to be the
generic way to send messages to stdout.  If the code were refactored to
actually use that, then it should be
straightforward to add an optional argument there